pyproject.toml
[project]
name = "agenticqueue-api"
version = "0.1.0"
description = "Migration tooling for the AgenticQueue API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "fastapi>=0.116.1,<1.0.0",
    "fastmcp>=2.0.0,<3.0.0",
    "httpx>=0.28.1,<1.0.0",
    "jsonschema>=4.25.1,<5.0.0",
    "mcp>=1.0.0,<2.0.0",
    "pgvector>=0.4.1,<1.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "psycopg[binary]>=3.2.9,<4.0.0",
    "pytest>=8.4.1,<9.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "pytest-timeout>=2.4.0,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "typer>=0.16.0,<1.0.0",
    "uvicorn>=0.35.0,<1.0.0",
    "argon2-cffi>=25.1.0,<26.0.0",
]

[project.scripts]
aq = "agenticqueue_cli.main:app"
aq-local = "agenticqueue_api.cli:app"
aq-mcp = "agenticqueue_cli.mcp:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps/api/src/agenticqueue_api", "apps/cli/src/agenticqueue_cli"]

[tool.pytest.ini_options]
pythonpath = ["apps/api/src", "apps/cli/src"]
testpaths = ["tests"]
addopts = "--cov=agenticqueue_api.middleware --cov-branch --cov-report=term-missing"

[tool.black]
target-version = ["py312"]
